some code review changes
