+++ server/data/nomis/prisonTimeline/PrisonApiPrisonPeriods.dto.ts
@@ -1,12 +1,9 @@
 export interface PrisonApiPrisonPeriodsDto {
   prisonerNumber: string
-  prisonPeriod:
-    | {
-        bookingNumber: string
-        bookingId: number
-        movementDates: {
-          dateInToPrison: string | undefined
-        }
-      }[]
-    | undefined
+  prisonPeriod?: {
+    bookingId: number
+    movementDates: {
+      dateInToPrison: string | undefined
+    }
+  }[]
 }
