+++ server/services/filter/homeFilter.ts
@@ -23,7 +23,7 @@ import { RisksAndNeedsApiClient } from '../../data/risksAndNeeds/risksAndNeedsAp
 import { OverallRiskLevel } from '../../data/risksAndNeeds/riskSummary.dto'
 import logger from '../../../log'
 import { RecalledOffenderData } from '../recategorisation/recall/recalledOffenderData'
-import { add10BusinessDays } from '../../utils/utilities'
+import utils from '../../utils/utils'
 
 export const SUITABILIGY_FOR_OPEN_CONDITIONS = 'suitabilityForOpenConditions'
 export const DUE_DATE = 'dueDate'
@@ -253,7 +253,7 @@ export const filterListOfPrisoners = async (
             currentPrisonerSearchData.recall &&
             recalledOffenderData.get(prisoner.offenderNo).recallDate
           ) {
-            nextReviewDate = add10BusinessDays(recalledOffenderData.get(prisoner.offenderNo).recallDate)
+            nextReviewDate = utils.add10BusinessDays(recalledOffenderData.get(prisoner.offenderNo).recallDate)
           }
           if (!isReviewOverdue(nextReviewDate)) {
             return false
